vc 项目文件，消除 vc 编译警告

+++ fibjs/src/Stats.cpp
@@ -166,7 +166,7 @@ result_t Stats::_named_enumerator(v8::Handle<v8::Array>& retVal)
 	retVal = v8::Array::New();
 
 	for (i = 0; i < m_size; i++)
-		retVal->Set(i, v8::String::New(m_keys[i].c_str(), m_keys[i].length()));
+		retVal->Set(i, v8::String::New(m_keys[i].c_str(), (int)m_keys[i].length()));
 
 	return 0;
 }
